app/requirements.txt
python-dateutil==2.8.1
python-dotenv==0.15.0